refactor: update project structure and README
- Moved logging configuration under the config folder
- Updated project structure in README to reflect the current folder organization
